fix(tests): fix property test failures - weighted median and overflow bugs
Fixed two critical bugs in property-based tests:

1. Weighted Median Algorithm Mismatch (oracle_properties_test.go)
   - Bug: Test implementation used `halfWeight = totalWeight / 2`
   - Fix: Changed to `halfWeight = (totalWeight + 1) / 2` to match production
   - This aligns with x/oracle/keeper/aggregation.go implementation
   - Added comment explaining the production alignment

2. Byzantine Resistance Test Flaw (oracle_properties_test.go)
   - Bug: Test assumed 45% attacker power can't control median with fragmented honest validators
   - Reality: Weighted median CAN select attacker price if honest validators below median have <50% cumulative weight
   - Fix: Changed test to verify attacker <50% can't control outcome when remaining validators AGREE on price
   - This is the actual Byzantine-resistant property guaranteed by weighted median

3. Integer Overflow in Resource Cost Calculation (compute_properties_test.go)
   - Bug: `sizeCost = uint64(requestSize + resultSize)` caused overflow when adding uint32 values
   - Fix: Changed to `sizeCost = uint64(requestSize) + uint64(resultSize)` to prevent overflow
   - Monotonicity property was failing on large uint32 inputs due to wraparound

Test Results:
- All property tests now pass consistently
- Tested with -race flag (no race conditions)
- Tested with -count=3 (stable across multiple runs)
- 34 property tests covering wallet, compute, DEX, and oracle modules

Security Impact:
- These were test-only bugs; production code was already correct
- Fixes ensure property tests accurately verify production invariants
- Overflow fix prevents false positives in resource cost monotonicity checks

+++ tests/property/compute_properties_test.go
@@ -359,7 +359,8 @@ func computeStateTransition(initialState, trace []byte) []byte {
 }
 
 func calculateResourceCost(requestSize, resultSize, execTime uint32) uint64 {
-	sizeCost := uint64(requestSize + resultSize)
+	// Convert to uint64 BEFORE adding to prevent overflow
+	sizeCost := uint64(requestSize) + uint64(resultSize)
 	timeCost := uint64(execTime) * 1000
 	return sizeCost + timeCost
 }
